--- src/main/java/site/mutopia/server/album/domain/AlbumTrack.java
@@ -1,9 +1,0 @@
-package site.mutopia.server.album.domain;
-
-public class AlbumTrack {
-    String id;
-    String name;
-    String order;
-    Long rating;
-    String length;
-}
