--- src/main/java/site/mutopia/server/album/domain/MutopiaAlbum.java
@@ -1,17 +1,0 @@
-package site.mutopia.server.album.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-@AllArgsConstructor
-public class MutopiaAlbum {
-    String id;
-    String name;
-    String artistName;
-    String coverImageUrl;
-    String releaseDate;
-    Long length;
-}
