--- src/main/java/site/mutopia/server/album/dto/AlbumTrackDto.java
@@ -1,10 +1,0 @@
-package site.mutopia.server.album.dto;
-
-public record AlbumTrackDto(
-        String id,
-        String name,
-        String order,
-        Long rating,
-        String length
-) {
-}
