--- src/main/java/site/mutopia/server/album/dto/RatingDistributionDto.java
@@ -1,7 +1,0 @@
-package site.mutopia.server.album.dto;
-
-public record RatingDistributionDto(
-        Double range,
-        Double ratio
-) {
-}
