--- src/main/java/site/mutopia/server/album/dto/TrendingAlbumInfoDto.java
@@ -1,9 +1,0 @@
-package site.mutopia.server.album.dto;
-
-public record TrendingAlbumInfoDto(
-        String albumId,
-        String albumName,
-        String albumImg,
-        String artistName
-) {
-}
