--- src/main/java/site/mutopia/server/album/dto/response/TrendingAlbumResDto.java
@@ -1,14 +1,0 @@
-package site.mutopia.server.album.dto.response;
-
-import site.mutopia.server.album.dto.TrendingAlbumInfoDto;
-
-import java.util.List;
-
-public record TrendingAlbumResDto(
-        List<TrendingAlbumInfoDto> albumList
-) {
-    /*
-    public TrendingAlbumResDto {
-        // TODO: Entity To Dto Mapper Logic
-    }*/
-}
