--- src/main/java/site/mutopia/server/album/service/AlbumService.java
@@ -1,28 +1,0 @@
-package site.mutopia.server.album.service;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import site.mutopia.server.album.domain.MutopiaAlbum;
-import site.mutopia.server.album.repository.AlbumRepository;
-
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-@Slf4j
-public class AlbumService {
-
-    private final AlbumRepository albumRepository;
-
-    public void findAlbumById(String albumId) {
-        MutopiaAlbum albumById = albumRepository.findAlbumById(albumId);
-        log.info("albumById: {}", albumById);
-    }
-
-    public void findAlbumByAlbumName(String albumName) {
-        List<MutopiaAlbum> albumByAlbumName = albumRepository.findAlbumByKeyword(albumName, 0, 0);
-        log.info("albumByAlbumName: {}", albumByAlbumName.stream().map(MutopiaAlbum::getName).toList());
-    }
-
-}
