--- src/main/java/site/mutopia/server/comment/controller/CommentController.java
@@ -1,4 +1,0 @@
-package site.mutopia.server.comment.controller;
-
-public class CommentController {
-}
