--- src/main/java/site/mutopia/server/config/SwaggerConfig.java
@@ -1,36 +1,0 @@
-package site.mutopia.server.config;
-
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.security.SecurityRequirement;
-import io.swagger.v3.oas.models.security.SecurityScheme;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-    @Bean
-    public OpenAPI openAPI() {
-        String jwt = "JWT";
-        SecurityRequirement securityRequirement = new SecurityRequirement().addList(jwt);
-        Components components = new Components()
-                .addSecuritySchemes(jwt, new SecurityScheme()
-                        .name(jwt)
-                        .type(SecurityScheme.Type.HTTP)
-                        .scheme("bearer")
-                        .bearerFormat(jwt));
-        return new OpenAPI()
-                .components(new Components()).info(apiInfo())
-                .addSecurityItem(securityRequirement)
-                .components(components);
-    }
-
-    private Info apiInfo() {
-        return new Info()
-                .title("Mutopia API")
-                .description("Mutopia API")
-                .version("1.0");
-    }
-
-}
