+++ src/main/java/site/mutopia/server/domain/album/domain/AlbumTrack.java
@@ -1,0 +1,9 @@
+package site.mutopia.server.domain.album.domain;
+
+public class AlbumTrack {
+    String id;
+    String name;
+    String order;
+    Long rating;
+    String length;
+}
