+++ src/main/java/site/mutopia/server/domain/album/dto/AlbumTrackDto.java
@@ -1,0 +1,10 @@
+package site.mutopia.server.domain.album.dto;
+
+public record AlbumTrackDto(
+        String id,
+        String name,
+        String order,
+        Long rating,
+        String length
+) {
+}
