+++ src/main/java/site/mutopia/server/domain/album/dto/RatingDistributionDto.java
@@ -1,0 +1,7 @@
+package site.mutopia.server.domain.album.dto;
+
+public record RatingDistributionDto(
+        Double range,
+        Double ratio
+) {
+}
