+++ src/main/java/site/mutopia/server/domain/album/dto/ReviewPreviewDto.java
@@ -1,0 +1,14 @@
+package site.mutopia.server.domain.album.dto;
+
+import java.time.ZonedDateTime;
+
+public record ReviewPreviewDto(
+        String reviewId,
+        String authorName,
+        String authorImg,
+        String title,
+        String content,
+        Long stars,
+        ZonedDateTime createdAt
+) {
+}
