+++ src/main/java/site/mutopia/server/domain/album/dto/TrendingAlbumInfoDto.java
@@ -1,0 +1,9 @@
+package site.mutopia.server.domain.album.dto;
+
+public record TrendingAlbumInfoDto(
+        String albumId,
+        String albumName,
+        String albumImg,
+        String artistName
+) {
+}
