+++ src/main/java/site/mutopia/server/domain/album/dto/response/TrendingAlbumResDto.java
@@ -1,0 +1,14 @@
+package site.mutopia.server.domain.album.dto.response;
+
+import site.mutopia.server.domain.album.dto.TrendingAlbumInfoDto;
+
+import java.util.List;
+
+public record TrendingAlbumResDto(
+        List<TrendingAlbumInfoDto> albumList
+) {
+    /*
+    public TrendingAlbumResDto {
+        // TODO: Entity To Dto Mapper Logic
+    }*/
+}
