+++ src/main/java/site/mutopia/server/domain/comment/controller/CommentController.java
@@ -1,0 +1,4 @@
+package site.mutopia.server.domain.comment.controller;
+
+public class CommentController {
+}
