+++ src/main/java/site/mutopia/server/domain/follow/controller/FollowController.java
@@ -1,0 +1,4 @@
+package site.mutopia.server.domain.follow.controller;
+
+public class FollowController {
+}
