+++ src/main/java/site/mutopia/server/domain/playlist/controller/PlaylistController.java
@@ -1,0 +1,4 @@
+package site.mutopia.server.domain.playlist.controller;
+
+public class PlaylistController {
+}
