+++ src/main/java/site/mutopia/server/domain/profile/controller/ProfileController.java
@@ -1,0 +1,14 @@
+package site.mutopia.server.domain.profile.controller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class ProfileController {
+
+    @GetMapping("/hi")
+    public String index() {
+        return "Hello";
+    }
+
+}
