+++ src/main/java/site/mutopia/server/domain/review/controller/ReviewController.java
@@ -1,0 +1,4 @@
+package site.mutopia.server.domain.review.controller;
+
+public class ReviewController {
+}
