--- src/main/java/site/mutopia/server/follow/controller/FollowController.java
@@ -1,4 +1,0 @@
-package site.mutopia.server.follow.controller;
-
-public class FollowController {
-}
