+++ src/main/java/site/mutopia/server/global/config/SecurityConfig.java
@@ -1,0 +1,24 @@
+package site.mutopia.server.global.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+@EnableWebSecurity
+public class SecurityConfig {
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+
+        http.csrf(AbstractHttpConfigurer::disable);
+
+        http.authorizeHttpRequests((auth) ->
+                auth.anyRequest().permitAll()
+        );
+
+        return http.build();
+    }
+}
