--- src/main/java/site/mutopia/server/playlist/controller/PlaylistController.java
@@ -1,4 +1,0 @@
-package site.mutopia.server.playlist.controller;
-
-public class PlaylistController {
-}
