--- src/main/java/site/mutopia/server/profile/controller/ProfileController.java
@@ -1,14 +1,0 @@
-package site.mutopia.server.profile.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class ProfileController {
-
-    @GetMapping("/hi")
-    public String index() {
-        return "Hello";
-    }
-
-}
