--- src/main/java/site/mutopia/server/review/controller/ReviewController.java
@@ -1,4 +1,0 @@
-package site.mutopia.server.review.controller;
-
-public class ReviewController {
-}
