--- src/main/java/site/mutopia/server/spotify/SpotifyConfig.java
@@ -1,34 +1,0 @@
-package site.mutopia.server.spotify;
-
-import org.apache.hc.core5.http.ParseException;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import se.michaelthelin.spotify.SpotifyApi;
-import se.michaelthelin.spotify.exceptions.SpotifyWebApiException;
-import se.michaelthelin.spotify.model_objects.credentials.ClientCredentials;
-
-import java.io.IOException;
-
-@Configuration
-public class SpotifyConfig {
-
-    @Value("${spotify.client.id}")
-    private String clientId;
-
-    @Value("${spotify.client.secret}")
-    private String clientSecret;
-
-    @Bean
-    public SpotifyApi spotifyApi() throws IOException, ParseException, SpotifyWebApiException {
-        SpotifyApi spotifyApi = new SpotifyApi.Builder()
-                .setClientId(clientId)
-                .setClientSecret(clientSecret)
-                .build();
-        ClientCredentials credentials = spotifyApi.clientCredentials().build().execute();
-        String accessToken = credentials.getAccessToken();
-        return new SpotifyApi.Builder()
-                .setAccessToken(accessToken)
-                .build();
-    }
-}
