+++ src/test/java/site/mutopia/server/album/service/AlbumServiceTest.java
@@ -3,6 +3,7 @@ package site.mutopia.server.album.service;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import site.mutopia.server.domain.album.service.AlbumService;
 
 @SpringBootTest
 class AlbumServiceTest {
